API reference

Helper types

COBREXA.Maybe — Type

Maybe{X}

Type of optional values.

Model loading and saving

COBREXA.download_model — Method

download_model(args...; kwargs...) -> Any

Safely download a model with a known hash. All arguments are forwarded to AbstractFBCModels.download_data_file – see the documentation in the AbstractFBCModels package for details.

COBREXA.load_model — Method

load_model(path::String) -> Any

Load a FBC model representation while guessing the correct model type. Uses AbstractFBCModels.load.

This overload almost always involves a search over types; do not use it in environments where performance is critical.

COBREXA.load_model — Method

load_model(
    model_type::Type{T<:AbstractFBCModels.AbstractFBCModel},
    path::String
) -> Any

Load a FBC model representation. Uses AbstractFBCModels.load.

COBREXA.save_model — Method

save_model(
    model::AbstractFBCModels.AbstractFBCModel,
    path::String
)

Save a FBC model representation. Uses AbstractFBCModels.save.

Solver interface

COBREXA.Feasible — Constant

Maximal

Objective sense for finding the any feasible value of the objective.

Same as JuMP.FEASIBILITY_SENSE.

COBREXA.Maximal — Constant

Maximal

Objective sense for finding the maximal value of the objective.

Same as JuMP.MAX_SENSE.

COBREXA.Minimal — Constant

Minimal

Objective sense for finding the minimal value of the objective.

Same as JuMP.MIN_SENSE.

COBREXA.Switch — Type

mutable struct Switch <: ConstraintTrees.Bound

Representation of a "binary switch" bound for ConstraintTrees. The value is constrained to be either the value of field a or of field b; both fields are Float64s. Upon translation to JuMP, the switches create an extra boolean variable, and the value is constrained to equal a + boolean_var * (b-a).

Switches can be offset by adding real numbers, negated, and multiplied and divided by scalar constraints. For optimizing some special cases, multiplying by exact zero returns an equality bound to zero.

Fields

a::Float64: One choice

b::Float64: The other choice

COBREXA.is_solved — Method

is_solved(opt_model::JuMP.Model) -> Bool

true if opt_model solved successfully (solution is optimal or locally optimal). false if any other termination status is reached.

COBREXA.optimization_model — Method

optimization_model(
    cs::Union{ConstraintTrees.Constraint, ConstraintTrees.Tree{ConstraintTrees.Constraint}};
    objective,
    optimizer,
    sense
)

Construct a JuMP Model that describes the precise constraint system into the JuMP Model created for solving in optimizer, with a given optional objective and optimization sense chosen from Maximal, Minimal and Feasible.

COBREXA.optimized_model — Method

optimized_model(om; output)

Like optimized_values, but works directly with a given JuMP model om without applying any settings or creating the optimization model.

To run the process manually, you can use optimization_model to convert the constraints into a suitable JuMP optimization model.

Constraint system building

Genetic constraints

Objective function helpers

COBREXA.squared_sum_error_value — Method

squared_sum_error_value(
    constraints::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    target
) -> Any

Construct a ConstraintTrees.Value out of squared error (in the RMSE-like squared-error sense) between the values in the constraint tree and the reference target.

target is a function that takes a symbol (key) and returns either a Float64 reference value, or nothing if the error of given key should not be considered.

COBREXA.squared_sum_value — Method

squared_sum_value(
    x::ConstraintTrees.Tree{ConstraintTrees.Constraint}
) -> Union{ConstraintTrees.LinearValue, ConstraintTrees.QuadraticValue}

Construct a ConstraintTrees.Value out of squared sum of all values directly present in a given constraint tree.

COBREXA.sum_value — Method

sum_value(x...) -> Any

Construct a ConstraintTrees.Value out of a sum of all values directly present in a given constraint tree.

Bounds&tolerances helpers

COBREXA.absolute_tolerance_bound — Method

absolute_tolerance_bound(tolerance) -> COBREXA.var"#191#192"

Make a function that returns absolute tolerance bounds, i.e. value - tolerance and value + tolerance in a tuple, in the increasing order.

COBREXA.relative_tolerance_bound — Method

relative_tolerance_bound(tolerance) -> COBREXA.var"#193#194"

Make a function that returns relative tolerance bounds, i.e. value / tolerance and value * tolerance in a tuple, in the increasing order.

Analysis functions

Analysis modifications

Distributed analysis