Front-end user interface

Flux balance analysis

COBREXA.flux_balance_analysis — Method

flux_balance_analysis(
    model::AbstractFBCModels.AbstractFBCModel;
    kwargs...
)

Compute an optimal objective-optimizing solution of the given model.

Most arguments are forwarded to optimized_values.

Returns a tree with the optimization solution of the same shape as given by flux_balance_constraints.

COBREXA.flux_balance_constraints — Method

flux_balance_constraints(
    model::AbstractFBCModels.AbstractFBCModel;
    interface,
    interface_name
) -> Any

A constraint tree that models the content of the given instance of AbstractFBCModel.

The constructed tree contains subtrees fluxes (with the reaction-defining "variables") and flux_stoichiometry (with the metabolite-balance-defining constraints), and a single constraint objective thad describes the objective function of the model.

Optionally if interface is specified, an "interface block" will be created within the constraint tree for later use as a "module" in creating bigger models (such as communities) using interface_constraints. The possible parameter values include:

nothing – default, no interface is created

:sbo – the interface gets created from model's SBO annotations)

:identifier_prefixes – the interface is guesstimated from commonly occurring adhoc reaction ID prefixes used in contemporary models

:boundary – the interface is created from all reactions that either only consume or only produce metabolites

Output interface name can be set via interface_name.

See Configuration for fine-tuning the default interface creation.

Flux variability analysis

COBREXA.flux_variability_analysis — Method

flux_variability_analysis(
    model::AbstractFBCModels.AbstractFBCModel;
    objective_bound,
    reaction_subset,
    optimizer,
    settings,
    workers
)

Perform a Flux Variability Analysis (FVA) on the model, and return a dictionary of flux ranges where the model is able to perform optimally.

The constraint system is constructed using flux_balance_constraints, and the variability is examined on all reaction's fluxes, or on the subset given optionally in reaction_subset (e.g., reaction_subset = ["PFK", "ACALD"]). The optimality tolerance can be specified with objectivebound using e.g. [`relativetolerancebound](@ref) or [absolutetolerance_bound`](@ref); the default is 99% relative tolerance.

Parameter workers may be used to enable parallel or distributed processing; the execution defaults to all available workers. Other parameters (esp. optimizer) are internally forwarded to optimized_values.

Use constraints_variability to customize the FVA execution.

Parsimonious flux balance analysis

COBREXA.linear_parsimonious_flux_balance_analysis — Method

linear_parsimonious_flux_balance_analysis(
    model::AbstractFBCModels.AbstractFBCModel;
    tolerances,
    kwargs...
)

Like parsimonious_flux_balance_analysis, but uses the L1-metric parsimonious system given by linear_parsimonious_flux_balance_constraints.

In turn, the solution is often faster, does not require a solver capable of quadratic objectives, and has many beneficial properties of the usual parsimonious solutions (such as the general lack of unnecessary loops). On the other hand, like with plain flux balance analysis there is no strong guarantee of uniqueness of the solution.

Solver configuration arguments are forwarded to parsimonious_optimized_values.

COBREXA.linear_parsimonious_flux_balance_constraints — Method

linear_parsimonious_flux_balance_constraints(
    model::AbstractFBCModels.AbstractFBCModel,
    optimizer;
    kwargs...
) -> Any

Like parsimonious_flux_balance_constraints, but uses a L1 metric for solving the parsimonious problem. The parsimonious_objective constraint is thus linear.

COBREXA.parsimonious_flux_balance_analysis — Method

parsimonious_flux_balance_analysis(
    model::AbstractFBCModels.AbstractFBCModel;
    tolerances,
    kwargs...
)

Compute a parsimonious flux solution for the model, using the constraints given by parsimonious_flux_balance_constraints.

In short, the objective value of the parsimonious solution should be the same as the one from flux_balance_analysis, except the squared sum of reaction fluxes is minimized. If there are multiple possible fluxes that achieve a given objective value, parsimonious flux thus represents the "minimum energy" one, which is arguably more realistic.

Solver configuration arguments are forwarded to parsimonious_optimized_values.

COBREXA.parsimonious_flux_balance_constraints — Method

parsimonious_flux_balance_constraints(
    model::AbstractFBCModels.AbstractFBCModel
) -> Any

A constraint system like from flux_balance_constraints, but with the parsimonious objective present under key parsimonious_objective. Best used via parsimonious_flux_balance_analysis.

Minimization of metabolic adjustment

COBREXA.linear_metabolic_adjustment_minimization_analysis — Method

linear_metabolic_adjustment_minimization_analysis(
    model::AbstractFBCModels.AbstractFBCModel,
    args...;
    optimizer,
    settings,
    reference_parsimonious_optimizer,
    reference_parsimonious_settings,
    reference_optimizer,
    reference_settings,
    kwargs...
)

TODO

COBREXA.linear_metabolic_adjustment_minimization_constraints — Method

linear_metabolic_adjustment_minimization_constraints(
    model::AbstractFBCModels.AbstractFBCModel,
    reference_model::AbstractFBCModels.AbstractFBCModel;
    kwargs...
) -> Union{Nothing, ConstraintTrees.Tree{ConstraintTrees.Constraint}}

Like metabolic_adjustment_minimization_constraints but optimizes the L1 norm. This typically produces a sufficiently good result with less resources, depending on the situation. See documentation of linear_parsimonious_flux_balance_analysis for some of the considerations.

COBREXA.linear_metabolic_adjustment_minimization_constraints — Method

linear_metabolic_adjustment_minimization_constraints(
    model::AbstractFBCModels.AbstractFBCModel,
    reference_fluxes::ConstraintTrees.Tree
) -> ConstraintTrees.Tree{ConstraintTrees.Constraint}

Like metabolic_adjustment_minimization_constraints but optimizes the L1 norm. This typically produces a sufficiently good result with less resources, depending on the situation. See documentation of linear_parsimonious_flux_balance_analysis for some of the considerations.

COBREXA.metabolic_adjustment_minimization_analysis — Method

metabolic_adjustment_minimization_analysis(
    model::AbstractFBCModels.AbstractFBCModel,
    args...;
    optimizer,
    settings,
    reference_parsimonious_optimizer,
    reference_parsimonious_settings,
    reference_optimizer,
    reference_settings,
    kwargs...
)

TODO

COBREXA.metabolic_adjustment_minimization_constraints — Method

metabolic_adjustment_minimization_constraints(
    model::AbstractFBCModels.AbstractFBCModel,
    reference_model::AbstractFBCModels.AbstractFBCModel;
    kwargs...
) -> Any

A slightly easier-to-use version of metabolic_adjustment_minimization_constraints that computes the reference flux as the parsimonious optimal solution of the reference_model. The reference flux is calculated using reference_optimizer and reference_modifications, which default to the optimizer and settings.

Other arguments are forwarded to the internal call of parsimonious_optimized_values.

Returns nothing if no feasible solution is found.

COBREXA.metabolic_adjustment_minimization_constraints — Method

metabolic_adjustment_minimization_constraints(
    model::AbstractFBCModels.AbstractFBCModel,
    reference_fluxes::ConstraintTrees.Tree
) -> Any

Find a solution of the "minimization of metabolic adjustment" (MOMA) analysis for the model, which is the "closest" feasible solution to the given reference_fluxes, in the sense of squared-sum distance. The minimized squared distance (the objective) is present in the result tree as minimal_adjustment_objective.

This is often used for models with smaller feasible region than the reference models (typically handicapped by a knockout, nutritional deficiency or a similar perturbation). MOMA solution then gives an expectable "easiest" adjustment of the organism towards a somewhat working state.

Reference fluxes that do not exist in the model are ignored (internally, the objective is constructed via squared_sum_error_value).

Additional parameters are forwarded to optimized_values.

Constraint systems for metabolite concentrations

COBREXA.log_concentration_constraints — Method

log_concentration_constraints(
    model::AbstractFBCModels.AbstractFBCModel;
    concentration_bound
) -> ConstraintTrees.Tree{ConstraintTrees.Constraint}

Build log-concentration-stoichiometry constraints for the model, as used e.g. by max_min_driving_force_analysis.

The output constraint tree contains a log-concentration variable for each metabolite in subtree log_concentrations. Individual reactions' total reactant log concentrations (i.e., all log concentrations of actual reactants minus all log concentrations of products) have their own variables in reactant_log_concentrations.

Function concentration_bound may return a bound for the log-concentration of a given metabolite (compatible with ConstraintTrees.Bound), or nothing.

Enzyme-mass-constrained models

COBREXA.Isozyme — Type

mutable struct Isozyme

A simple struct storing information about the isozyme composition, including subunit stoichiometry and turnover numbers. Use with enzyme_constrained_flux_balance_analysis.

Fields

gene_product_stoichiometry::Dict{String, Float64}

kcat_forward::Union{Nothing, Float64}

kcat_reverse::Union{Nothing, Float64}

COBREXA.enzyme_constrained_flux_balance_analysis — Method

enzyme_constrained_flux_balance_analysis(
    model::AbstractFBCModels.AbstractFBCModel;
    kwargs...
)

Perform the enzyme-constrained flux balance analysis on the model and return the solved constraint system.

Arguments are forwarded to enzyme_constrained_flux_balance_constraints; solver configuration arguments are forwarded to optimized_values.

COBREXA.enzyme_constrained_flux_balance_constraints — Method

enzyme_constrained_flux_balance_constraints(
    model::AbstractFBCModels.AbstractFBCModel;
    reaction_isozymes,
    gene_product_molar_masses,
    capacity
)

Construct a enzyme-constrained flux-balance constraint system. The model is parameterized by reaction_isozymes, which is a mapping of reaction identifiers to Isozyme descriptions.

Additionally, the computation requires gene_product_molar_masses to describe the weights of enzyme building material, and capacity, which limits the mass of enzymes in the whole model.

capacity may be a single number, which sets the limit for "all described enzymes". Alternatively, capacity may be a vector of identifier-genes-limit triples that make a constraint (identified by the given identifier) that limits the listed genes to the given limit.

Community models

Production envelopes

COBREXA.objective_production_envelope — Method

objective_production_envelope(
    model::AbstractFBCModels.AbstractFBCModel,
    reactions::Vector{String};
    breaks,
    optimizer,
    settings,
    workers
)

Find the objective production envelope of the model in the dimensions given by reactions.

This runs a variability analysis of the fluxes given by flux_balance_constraints to determine an applicable range for the dimensions, then splits the dimensions to equal-sized breaks (of count breaks for each dimension, i.e. total breaks ^ length(reactions) individual "multidimensional breaks") thus forming a grid, and returns an array of fluxes through the model objective with the individual reactions fixed to flux as given by the grid.

optimizer and settings are used to construct the optimization models.

The computation is distributed to the specified workers, defaulting to all available workers.

Knockout models

Loopless flux balance analysis

COBREXA.loopless_flux_balance_analysis — Method

loopless_flux_balance_analysis(
    model::AbstractFBCModels.AbstractFBCModel;
    kwargs...
)

Perform the loopless flux balance analysis on the model, returning the solved constraint system.

Arguments are forwarded to loopless_flux_balance_constraints (see the documentation for the description of the constraint system); solver configuration arguments are forwarded to optimized_values.

COBREXA.loopless_flux_balance_constraints — Method

loopless_flux_balance_constraints(
    model::AbstractFBCModels.AbstractFBCModel;
    flux_infinity_bound,
    driving_force_nonzero_bound,
    driving_force_infinity_bound
) -> ConstraintTrees.Tree{ConstraintTrees.Constraint}

Construct a flux-balance constraint system from model with added quasi-thermodynamic constraints that ensure that thermodynamically infeasible internal cycles can not occur. The method is closer described by: Schellenberger, Lewis, and, Palsson. "Elimination of thermodynamically infeasible loops in steady-state metabolic models.", Biophysical journal, 2011`.

The loopless condition comes with a performance overhead: the computation needs to find the null space of the stoichiometry matrix (essentially inverting it); and the subsequently created optimization problem contains binary variables for each internal reaction, thus requiring a MILP solver and a potentially exponential solving time.

Internally, the system is constructed by combining flux_balance_constraints and loopless_constraints.

The arguments driving_force_max_bound and driving_force_nonzero_bound set the bounds (possibly negated ones) on the virtual "driving forces" (G_i in the paper).

Max-Min Driving Force analysis

COBREXA.max_min_driving_force_analysis — Method

max_min_driving_force_analysis(
    model::AbstractFBCModels.AbstractFBCModel;
    kwargs...
)

Perform the max-min driving force analysis on the model, returning the solved constraint system.

Arguments are forwarded to max_min_driving_force_constraints (see the documentation for the description of the constraint system); solver configuration arguments are forwarded to optimized_values.

COBREXA.max_min_driving_force_constraints — Method

max_min_driving_force_constraints(
    model::AbstractFBCModels.AbstractFBCModel;
    reaction_standard_gibbs_free_energies,
    reference_flux,
    concentration_ratios,
    constant_concentrations,
    ignored_metabolites,
    proton_metabolites,
    water_metabolites,
    concentration_lower_bound,
    concentration_upper_bound,
    T,
    R,
    reference_flux_atol,
    check_ignored_reactions
)

Create max-min driving force constraint system from model, using the supplied reaction standard Gibbs free energies in reaction_standard_gibbs_free_energies.

The method is described by: Noor, et al., "Pathway thermodynamics highlights kinetic obstacles in central metabolism.", PLoS computational biology, 2014.

reference_flux sets the directions of each reaction in model. The scale of the values is not important, only the direction is examined (w.r.t. reference_flux_atol tolerance). Ideally, the supplied reference_flux should be completely free of internal cycles, which enables the thermodynamic consistency. To get the cycle-free flux, you can use loopless_flux_balance_analysis (computationally demanding, but gives thermodynamically consistent solutions), parsimonious_flux_balance_analysis or linear_parsimonious_flux_balance_analysis (which is computationally simple, but the consistency is not guaranteed).

Internally, log_concentration_constraints is used to lay out the base structure of the problem.

Following arguments are set optionally:

water_metabolites, proton_metabolites and ignored_metabolites allow to completely ignore constraints on a part of the metabolite set, which is explicitly recommended especially for water and protons (since the analyses generally assume aqueous environment of constant pH)

constant_concentrations can be used to fix the concentrations of the metabolites

concentration_lower_bound and concentration_upper_bound set the default concentration bounds for all other metabolites

concentration ratios is a dictionary that assigns a tuple of metabolite-metabolite-concentration ratio constraint names; e.g. ATP/ADP ratio can be fixed to five-times-more-ATP by setting concentration_ratios = Dict("adenosin_ratio" => ("atp", "adp", 5.0))

T and R default to the usual required thermodynamic constraints in the usual units (K and kJ/K/mol, respectively). These multiply the log-concentrations to obtain the actual Gibbs energies and thus driving forces.

Sampling

COBREXA.achr_flux_sample — Method

achr_flux_sample(
    model::AbstractFBCModels.AbstractFBCModel;
    optimizer,
    kwargs...
)

TODO

COBREXA.affine_hr_flux_sample — Method

affine_hr_flux_sample(
    model::AbstractFBCModels.AbstractFBCModel;
    optimizer,
    kwargs...
)

TODO