Miscellaneous functions

COBREXA.absolute_tolerance_bound — Method

absolute_tolerance_bound(tolerance) -> COBREXA.var"#263#264"

Make a function that returns absolute tolerance bounds, i.e. value - tolerance and value + tolerance in a tuple, in the increasing order.

COBREXA.relative_tolerance_bound — Method

relative_tolerance_bound(tolerance) -> COBREXA.var"#265#266"

Make a function that returns relative tolerance bounds, i.e. value / tolerance and value * tolerance in a tuple, in the increasing order.

COBREXA.break_interval — Method

break_interval(lower, upper, breaks::Int64) -> Any

Break an interval into breaks (count) breaks.

Used for computing breaks in objective_production_envelope.

COBREXA.maybeget — Method

maybeget(_::Nothing, _...)

Helper for getting stuff from dictionaries where keys may be easily missing.

COBREXA.maybemap — Method

maybemap(f, _::Nothing)

Helper for applying functions to stuff that might be nothing.

COBREXA.set_objective_sense — Method

set_objective_sense(
    objective_sense
) -> COBREXA.var"#267#268"

Change the objective sense of optimization. Accepted arguments include Minimal, Maximal, and Feasible.

COBREXA.set_optimizer — Method

set_optimizer(optimizer) -> COBREXA.var"#269#270"

Change the JuMP optimizer used to run the optimization.

COBREXA.set_optimizer_attribute — Method

set_optimizer_attribute(
    attribute_key,
    value
) -> COBREXA.var"#271#272"

Change a JuMP optimizer attribute. The attributes are optimizer-specific, refer to the JuMP documentation and the documentation of the specific optimizer for usable keys and values.

COBREXA.set_time_limit_sec — Method

set_time_limit_sec(limit) -> COBREXA.var"#273#274"

Portable way to set a time limit in seconds for the optimizer computation.

COBREXA.silence — Method

silence

Modification that disable all output from the JuMP optimizer (shortcut for set_silent from JuMP).

COBREXA.tree_deflate — Method

tree_deflate(f, x::ConstraintTrees.Tree{T}) -> Vector

Extract all elements of a ConstraintTrees.Tree in order and return them in a Vector transformed by f. If the order is not modified, one can re-insert a vector of modified elements into the same-shaped tree using tree_reinflate.

COBREXA.tree_reinflate — Method

tree_reinflate(
    x::ConstraintTrees.Tree,
    elems::Array{T, 1}
) -> ConstraintTrees.Tree

Insert a Vector of elements into the "values" of a ConstraintTrees.Tree. The order of elements is given by tree_deflate.