Constraint system builders

Generic constraints

Comparisons

COBREXA.all_equal_constraints — Method

all_equal_constraints(
    a,
    tree::ConstraintTrees.Tree{ConstraintTrees.Constraint}
) -> Any

A constriant tree that makes sure that all values in tree are the same as the value of a.

Names in the output ConstraintTree match the names in the tree.

COBREXA.difference_constraint — Method

difference_constraint(
    a,
    b,
    difference_bound
) -> ConstraintTrees.Constraint

A constraint that makes sure that the difference from a to b is within the difference_bound. For example, difference_constraint(-1, 1, 2) will always be valid. Any type of ConstraintTree.Bound can be supplied.

COBREXA.equal_value_constraint — Method

equal_value_constraint(a, b) -> ConstraintTrees.Constraint

A constraint that makes sure that the values of a and b are the same.

COBREXA.greater_or_equal_constraint — Method

greater_or_equal_constraint(
    a,
    b
) -> ConstraintTrees.Constraint

A constraint that makes sure that the value of a is greater than or equal to the the value of b.

COBREXA.less_or_equal_constraint — Method

less_or_equal_constraint(a, b) -> ConstraintTrees.Constraint

A constraint that makes sure that the value of a is less than or equal to the the value of b.

Constraint scaling

COBREXA.scale_bounds — Method

scale_bounds(
    tree::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    factor
) -> Any

Linearly scale all bounds in a constraint tree by the factor. This actually changes the model semantics, and may not work in surprising/improper ways with some constraint systems, esp. the MILP and QP ones.

See also scale_constraints.

COBREXA.scale_constraints — Method

scale_constraints(
    tree::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    factor
) -> Any

Linearly scale all constraints in a constraint tree by the factor.

See also scale_bounds.

Sign splitting

COBREXA.positive_bound_contribution — Method

positive_bound_contribution(
    b::ConstraintTrees.EqualTo
) -> ConstraintTrees.EqualTo

TODO

COBREXA.sign_split_constraints — Method

sign_split_constraints(; positive, negative, signed)

A constraint tree that bound the values present in signed to be sums of pairs of positive and negative contributions to the individual values.

Keys in the result are the same as the keys of signed constraints.

Typically, this can be used to create "unidirectional" fluxes together with unsigned_negative_contribution_variables and unsigned_positive_contribution_variables.

Use sign_split_variables to allocate the variables easily.

COBREXA.sign_split_variables — Method

sign_split_variables(
    constraints::ConstraintTrees.Tree{ConstraintTrees.Constraint};
    positive,
    negative
)

Shortcut for making a pair of named variable groups created by unsigned_positive_contribution_variables and unsigned_negative_contribution_variables, in subtrees named by positive and negative.

Use sign_split_constraints to bind the new variables to existing values.

COBREXA.unsigned_negative_contribution_variables — Method

unsigned_negative_contribution_variables(
    cs::ConstraintTrees.Tree{ConstraintTrees.Constraint}
) -> Any

TODO

COBREXA.unsigned_positive_contribution_variables — Method

unsigned_positive_contribution_variables(
    cs::ConstraintTrees.Tree{ConstraintTrees.Constraint}
) -> Any

TODO

Objectives

COBREXA.squared_sum_error_value — Method

squared_sum_error_value(
    constraints::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    target
) -> Any

Construct a ConstraintTrees.Value out of squared error (in the RMSE-like squared-error sense) between the values in the constraint tree and the reference target.

target is a function that takes a symbol (key) and returns either a Float64 reference value, or nothing if the error of given key should not be considered.

COBREXA.squared_sum_value — Method

squared_sum_value(
    x::ConstraintTrees.Tree{ConstraintTrees.Constraint}
) -> Union{ConstraintTrees.LinearValue, ConstraintTrees.QuadraticValue}

Construct a ConstraintTrees.Value out of squared sum of all values directly present in a given constraint tree.

COBREXA.sum_value — Method

sum_value(x...) -> Any

Construct a ConstraintTrees.Value out of a sum of all values directly present in a given constraint tree.

FBC constraint systems

Analysis-specific constriants

COBREXA.loopless_constraints — Method

loopless_constraints(
;
    fluxes,
    loopless_direction_indicators,
    loopless_driving_forces,
    internal_reactions,
    internal_nullspace,
    flux_infinity_bound,
    driving_force_nonzero_bound,
    driving_force_infinity_bound
)

Construct the loopless constraint system that binds fluxes of all internal_reactions to direction of loopless_direction_indicators and connects them to loopless_driving_forces. The solution is bounded to lie in internal_nullspace (which is a sufficient algebraic condition for loop-less-ness).

The indicators must be discrete variables, valued 1 if the reaction flux goes forward, or 0 if the reaction flux is reversed.

The simplest (but by no means the fastest) way to obtain a good internal_nullspace is to use LinearAlgebra.nullspace with the internal reactions' stoichiometry matrix. Rows of internal_nullspace must correspond to internal_reactions.

flux_infinity_bound is used as the maximal bound for fluxes (for constraints that connect them to indicator variables); it should optimally be greater than the maximum possible absolute value of any flux in the original model.

driving_force_nonzero_bound and driving_force_infinity_bound are similarly used to limit the individual reaction's driving forces.

COBREXA.fbc_gene_knockout_constraints — Method

fbc_gene_knockout_constraints(
;
    fluxes,
    knockout_genes,
    model
)

Make a ConstraintTree that simulates a gene knockout of knockout_genes in the model and disables corresponding fluxes accordingly.

Keys of the fluxes must correspond to the reaction identifiers in the model.

knockout_genes may be any collection that support element tests using in. Since the test is done many times, a Set is a preferred contained for longer lists of genes.

COBREXA.knockout_constraints — Method

knockout_constraints(
    knockout_test,
    fluxes::ConstraintTrees.Tree{ConstraintTrees.Constraint}
) -> ConstraintTrees.Tree{ConstraintTrees.Constraint}

Make a ConstraintTree that knocks out fluxes given by the predicate knockout_test. The predicate function is called with a single parameter (the key of the flux in tree fluxes) and must return a boolean. Returning true means that the corresponding flux (usually a reaction flux) will be knocked out.

Use fbc_gene_knockout_constraints to apply gene knockouts easily to models with AbstractFBCModel interface.

COBREXA.gene_product_isozyme_constraints — Method

gene_product_isozyme_constraints(
    gene_product_amounts::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    isozymes_amount_trees,
    isozyme_stoichiometry
) -> ConstraintTrees.Tree{ConstraintTrees.Constraint}

A constraint tree that binds the isozyme amounts to gene product amounts accordingly to their multiplicities (aka. stoichiometries, protein units, ...) given by isozyme_stoichiometry.

Values in gene_product_amounts should describe the gene product allocations.

isozyme_amount_trees is an iterable that contains ConstraintTrees that describe the allocated isozyme amounts (such as created by isozyme_amount_variables. One may typically pass in both forward- and reverse-direction amounts at once, but it is possible to use a single tree, e.g., in a uni-tuple: tuple(my_tree).

isozyme_stoichiometry gets called with a reaction and isozyme ID as given by the isozyme amount trees. It may return nothing in case there's no information.

COBREXA.isozyme_amount_variables — Method

isozyme_amount_variables(fluxes, flux_isozymes) -> Any

Create a ConstraintTree with variables for isozyme contributions to reaction fluxes. The tree has 2 levels: the first contains all reaction flux IDs that have isozymes, the second contains the isozyme IDs for each reaction flux.

fluxes should be anything that can be iterated to give reaction flux IDs.

flux_isozymes is a function that, for a given reaction flux ID, returns anything iterable that contains the isozyme IDs for the given reaction flux. Returning an empty iterable prevents allocating the subtree for the given flux.

COBREXA.isozyme_flux_constraints — Method

isozyme_flux_constraints(
    isozyme_amounts::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    fluxes::ConstraintTrees.Tree{ConstraintTrees.Constraint},
    kcat
) -> ConstraintTrees.Tree{ConstraintTrees.Constraint}

A constraint tree that sums up partial contributions of reaction isozymes to the fluxes of reactions.

For practical purposes, both fluxes and isozymes are here considered to be unidirectional, i.e., one would typically apply this twice to constraint both "forward" and "reverse" fluxes.

Function kcat should retugn the kcat value for a given reaction and isozyme (IDs of which respectively form the 2 parameters for each call).

Interfacing of constraint systems

COBREXA.interface_constraints — Method

interface_constraints(
    kv;
    kwargs...
) -> ConstraintTrees.Tree{ConstraintTrees.Constraint}

Overload of interface_constraints for general key-value containers.

COBREXA.interface_constraints — Method

interface_constraints(
    ps::Pair...;
    default_interface,
    out_interface,
    out_balance,
    ignore,
    bound
) -> ConstraintTrees.Tree{ConstraintTrees.Constraint}

Join multiple constraint tree modules with interfaces into a bigger module with an interface.

Modules are like usual constraint trees, but contain an explicitly declared interface part, marked properly in arguments using e.g. a tuple (the parameters should form a dictionary constructor that would generally look such as :module_name => (module, module.interface); the second tuple member may also be specified just by name as e.g. :interface, or omitted while relying on default_interface).

Interface parts get merged and constrained to create a new interface; networks are intact with disjoint variable sets.

Compatible modules with ready-made interfaces may be created e.g. by flux_balance_constraints.

ignore may be used to selectively ignore parts of interfaces given the "module name" identifier and constraint path in the interface (these form 2 parameters passed to ignore). Similarly, bound may be used to specify bounds for the new interface, if required.